GetARGB32() can take normal ints instead.

+++ ps3_main.cpp
@@ -94,7 +94,7 @@ int main(int argc,const char *argv[])
   return 0;
 }
 
-u32 GetARGB32(u16 r, u16 g, u16 b)
+u32 GetARGB32(int r, int g, int b)
 {
     u32 result = 0x00000000;
     result += (u32)((r << 16));
